Split tests into CopyableThrowable and constructor-related ones
