+++ kotlinx-coroutines-core/jvm/test/exceptions/StackTraceRecoveryExceptionConstructorsTest.kt
@@ -2,54 +2,24 @@ package kotlinx.coroutines.exceptions
 
 import kotlinx.coroutines.testing.*
 import kotlinx.coroutines.*
-import kotlinx.coroutines.channels.*
-import org.junit.Test
+import kotlinx.coroutines.channels.produce
 import kotlin.test.*
 
-@Suppress("UNREACHABLE_CODE", "UNUSED", "UNUSED_PARAMETER")
-class StackTraceRecoveryCustomExceptionsTest : TestBase() {
-
+class StackTraceRecoveryExceptionConstructorsTest: TestBase() {
     internal class NonCopyable(val customData: Int) : Throwable() {
         // Bait
         constructor(cause: Throwable) : this(42)
     }
 
-    internal class Copyable(val customData: Int) : Throwable(), CopyableThrowable<Copyable> {
-        // Bait
-        constructor(cause: Throwable) : this(42)
-
-        override fun createCopy(): Copyable {
-            val copy = Copyable(customData)
-            copy.initCause(this)
-            return copy
-        }
-    }
-
     @Test
     fun testStackTraceNotRecovered() = runTest {
+        val exception = NonCopyable(239)
         try {
             withContext(wrapperDispatcher(coroutineContext)) {
-                throw NonCopyable(239)
+                throw exception
             }
-            expectUnreached()
         } catch (e: NonCopyable) {
-            assertEquals(239, e.customData)
-            assertNull(e.cause)
-        }
-    }
-
-    @Test
-    fun testStackTraceRecovered() = runTest {
-        try {
-            withContext(wrapperDispatcher(coroutineContext)) {
-                throw Copyable(239)
-            }
-            expectUnreached()
-        } catch (e: Copyable) {
-            assertEquals(239, e.customData)
-            val cause = e.cause
-            assertIs<Copyable>(cause)
-            assertEquals(239, cause.customData)
+            assertSame(exception, e)
         }
     }
 
@@ -61,7 +31,6 @@ class StackTraceRecoveryCustomExceptionsTest : TestBase() {
             withContext(wrapperDispatcher(coroutineContext)) {
                 throw WithDefault("custom")
             }
-            expectUnreached()
         } catch (e: WithDefault) {
             assertEquals("custom", e.message)
             val cause = e.cause
@@ -83,6 +52,7 @@ class StackTraceRecoveryCustomExceptionsTest : TestBase() {
         assertIs<WrongMessageException>(ex)
         assertEquals("Token OK", ex.message)
     }
+
 
     @Test
     fun testNestedExceptionWithCause() = runTest {
@@ -117,46 +87,5 @@ class StackTraceRecoveryCustomExceptionsTest : TestBase() {
         }.exceptionOrNull() ?: error("Expected to fail")
         assertIs<WrongMessageException>(ex)
         assertEquals("Token OK", ex.message)
-    }
-
-    class CopyableWithCustomMessage(
-        message: String?,
-        cause: Throwable? = null
-    ) : RuntimeException(message, cause),
-        CopyableThrowable<CopyableWithCustomMessage> {
-
-        override fun createCopy(): CopyableWithCustomMessage {
-            return CopyableWithCustomMessage("Recovered: [$message]", cause)
-        }
-    }
-
-    @Test
-    fun testCustomCopyableMessage() = runTest {
-        val result = runCatching {
-            coroutineScope<Unit> {
-                throw CopyableWithCustomMessage("OK")
-            }
-        }
-        val ex = result.exceptionOrNull() ?: error("Expected to fail")
-        assertIs<CopyableWithCustomMessage>(ex)
-        assertEquals("Recovered: [OK]", ex.message)
-    }
-
-    @Test
-    fun testTryCopyThrows() = runTest {
-        class FailingException : Exception(), CopyableThrowable<FailingException> {
-            override fun createCopy(): FailingException? {
-                TODO("Not yet implemented")
-            }
-        }
-
-        val e = FailingException()
-        val result = runCatching {
-            coroutineScope<Unit> {
-                throw e
-            }
-        }
-
-        assertSame(e, result.exceptionOrNull())
     }
 }
